* | router complete replaced
